Cleanup, button shadow and some minor stuff
