--- ScreenshotFactory/Constants.swift
@@ -1,18 +1,0 @@
-//
-//  Constants.swift
-//  ScreenshotFactory
-//
-//  Created by Filip Němeček on 10/09/2019.
-//  Copyright © 2019 Filip Němeček. All rights reserved.
-//
-
-import Foundation
-
-enum R {
-    struct Images {
-        struct Frames {
-            static let iPhoneX = "iPhoneXS_Max_frame.png"
-        }
-        static let demoScreenshot = "demo-screenshot.png"
-    }
-}
